+++ src/main/java/com/cjm721/overloaded/item/functional/ItemMultiTool.java
@@ -478,11 +478,6 @@ public class ItemMultiTool extends PowerModItem {
   public static class ClientSideEvents {
     @SubscribeEvent
     public static void leftClickBlock(@Nonnull PlayerInteractEvent.LeftClickBlock event) {
-      // TODO This event is not firing on client currently. Letting run on server in single player
-      // so still kinda works
-      //      if(event.getSide() == LogicalSide.SERVER || event.getEntityPlayer() != Minecraft.getInstance().player)
-      //        return;
-
       if (!event
           .getEntityPlayer()
           .getUniqueID()
@@ -529,23 +524,6 @@ public class ItemMultiTool extends PowerModItem {
         }
       }
       event.getDrops().clear();
-    }
-  }
-
-  @Mod.EventBusSubscriber(
-      modid = MODID,
-      bus = Mod.EventBusSubscriber.Bus.FORGE,
-      value = Dist.DEDICATED_SERVER)
-  public static class DedicatedServerSideEvent {
-    @SubscribeEvent
-    public static void leftClickBlock(@Nonnull PlayerInteractEvent.LeftClickBlock event) {
-      // TODO: Remove. Only have as a workaround until
-      ItemStack stack = event.getItemStack();
-      if (stack.getItem().equals(ModItems.multiTool)) {
-        leftClickOnBlockServer(
-            (ServerPlayerEntity) event.getEntityPlayer(),
-            new LeftClickBlockMessage(event.getPos()));
-      }
     }
   }
 }